+++ TablePro/Models/Connection/DatabaseConnection.swift
@@ -237,6 +237,7 @@ extension DatabaseType {
     static let scylladb = DatabaseType(rawValue: "ScyllaDB")
     static let etcd = DatabaseType(rawValue: "etcd")
     static let cloudflareD1 = DatabaseType(rawValue: "Cloudflare D1")
+    static let dynamodb = DatabaseType(rawValue: "DynamoDB")
 }
 
 extension DatabaseType: Codable {